+++ cli/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Updated `files_artifact_mounting` internal test in order to test duplicate mountpoints
+
 ### Breaking Changes
 * Upgraded `kurtosis engine lib` to 1.15.0
   * Users should upgrade to `kurtosis-engine-api-lib` 1.15.0